Add base ctrl, form and pages ahead of first page
We're copying the Waste carriers pattern for a journey, using AASM to manage the state. The WCR pattern uses form objects to handle validations, and has a controller per page.

Both the forms and the controllers inherit from a base, and have default actions like going to an invalid page in the event an ID is unrecognised.

The code was essentially copy and pasted from waste carriers renewals, and is not as clean as it should be. However we are just trying to get the model up and running with Waste Exemptions, which for example uses ActiveRecord instead of Mongoid, and does not assume a logged in user.

We've also switched reg_identifier for ID for now, but will select a more appropriate reference once we have a working partial section of the journey.