Add ability 2 have migrations automatically loaded
The rails guide expects you to copy the migrations from the engine into the host app using `rails blorgh:install:migrations`.

This might be fine for a released and reusable engine like Devise, but in our case the engine is the heart of the app, and the migrations are likely to change frequently. Therefore we want to keep the migrations in the engine, and just have the app see and include them when migrations are run.

This change does this. Based on https://content.pivotal.io/blog/leave-your-migrations-in-your-rails-engines