Add start page (first) to the journey
This change adds the first page of the journey to the engine, based on the Waste Carriers model (and the code has essentially been copy and pasted over).

Unlike the first page in the waste carriers renewals journey

- the user is not logged in
- the user needs to select an option
- we are not trying to match against an existing registration

We also have the issue that we are dealing with active record, and not mongoid, and we have not defined what we will be holding to track the enrollment between pages, hence we are still sticking with ID.

However with these changes, we can present a page, validate that an option as been selected, and when submitted navigate to the next page (which still needs to be created).

One know issue at this point is that if a validation error is presented, the radio buttons disappear. However we want to focus on the journey and the state machine at this point so will pick this up later.