Add pages for waste operation location question
This change adds the main location question page, plus the end pages should an option other than England be selected.

It's been copied from waste carriers where the context is about the principal place of business. In waste exemptions however, its about the location or site of the waste operation.

Other changes are that we do not expect or support overseas users hence the option is removed. And unlike waste carriers you cannot still continue if you select an option other than England.