Fix broken translations for invalid page
The content and use of the invalid page is essentially broken. We have copied it from waste carriers where its most appropriately used when someone tries to access the renewals journey using an invalid or unrecognised registration number.

Currently we are just using the record ID between pages whilst we get the journey implemented, though this will change as it will mean anyone could simply change the ID and possibly see someone elses exemption details.

However until we implement a solution for that, we are using ID, and should that be invalid we will end up here. So for now with these changes it at least doesn't have missing translations.