Ensure migrations are consistent
This change ensures we keep to conventions for the naming of the migrations, by adding ToTablename e.g. ToEnrollments as a postfix to each migrations name.

Doing this in combination with the generator (e.g. `be rails generate migration AddContactName first_name:string last_name:string`) will ensure rails can create the migration with the right table name, and keep our migrations consistent.