Add contact the env agency page
If the user selects the 'change' option on the start page then they need to be redirected to the Contact the Environment Agency page.

Initially we did not save the value of what was selected on the start page, but to keep consistent with the current workflow design and to use the state machine to manage the transition we have had to add the attribute to the Enrollment.

Hence there are additional changes around the start forms.