Add token to enrollment and use instead of ID
Because we do not require a user to login before submitting an enrollment we can't use the standard authentication mechanisms to ensure they are permitted to work with a specific exemption.

As the user moves between the pages we need some way to tell the app to retrieve and display the details of the enrollment they are working with.

We could use ID, but as that is simply a sequential number, a malicious user could try changing the ID then submitting a GET request in the hope of viewing another users enrollment.

Though we believe this is a highly unlikely scenario, we have chosen to generate a 24-character unique token for each enrollment when initialised. We then use this to identify the enrollment in progress. We believe there is no chance a malicious user could guess the 24 character token of another enrollment in progress, so this issue is resolved.

Unfortunately replacing ID for token meant changing a large number of files, however the change in itself is relatively minor. Also now we have a fixed uniqiue identifier we have been able to implement a proper validator for it.