Add rubocop support to the project and resolve
We had previously added the config file for rubocop, but had omitted to add the gem to the gemspec to support its use in our CI pipeline.

This change adds the gem to the project, corrects the config to ensure the dummy app is fully excluded, and resolves any remaining violations.