Add registration no page + comp. house support
This change adds the company registration number page. This is only shown to those who select either limited company or LLP as their business type. They are required to enter their companies house registration number.

In the previous version of the service the only validation performed was on the format of the number. It never actually validated the reference against companies house.

The changes here not only bring over the WCR registration number page, but also its companies house validation and lookup service. So going forward the WEX service will also be able to validate that a registration not only exists, but belongs to an active company.