Add contact position field (#5)
This change follows the existing pattern for adding pages, though there is a slight change in its validator. Position is an optional field in the waste exemptions service and as such we have added a new position validator that permits a blank field.

If its not blank then we validate its not too long and doesn't contain any characters we would not expect in a position.