Add is a farm page
This deviates slightly from the pattern set down in Waste Carriers Renewals in that rather than holding a string value for the answer (either "yes" or "no") we are able to do what we should of done which is hold the result as a boolean. The field permits null and this represents an unanswered result.

Other than that it implements the WCR pattern for a Yes/No page.