Add operator address lookup selection page (#11)
Following on from adding the first postcode lookup page, we can now add the address selection page.

Again we have taken the WCR code as a basis, however we have had to make a number of changes. First, the results we get back are in a slightly different format because we are using the addressbase facade in WEX. Secondly we are dealing with active record associations.

We have also needed to add another model for holding the address, and linked it to enrollment via a has_many association.

Other than that though, the implementation follows a standard model controller, form objects, validation and locales.