Add operator manual address entry page
With the operator postcode and address lookup pages sorted, the next step was to bring in the manual address entry page. Like the previous changes this brings in the base classes and additional migrations to support this work, so when we come to add these address pages for other areas the change will be much smaller.

We've also had to make some alterations based on the fact that we are working with a PostgresSql backed model, the most impacted being mass attribute assignment. Rails has built in protections to prevent you from mass assignining attributes based on params the user has entered, so we have had to incorporate this within the changes made.