Refactor address logic in engine
We break down the address logic into 3 areas

- postcode lookup
- selecting a lookup result
- entering an address manually

When migrating the code from WCR we observed that there was a lot of similarities and duplication in the last 2 of those areas.

So before we duplicate this even further in WEX (as we have 2 other addresses to capture) we are taking the opportunity to remove the duplication we can, and make consistent the stuff we cannot as much as possible.