Update address lookup to use just postcode attr
To make the operator address lookup form consistent with the operator address manual form this change renames the `operator_postcode` attribute to simply `postcode`.

It means we can also bring the attribute down to the `address_lookup` object rather than repeating it in all subsequent lookup form objects we'll create for the other addresses.