Move base intialization into the base address form
Prior to this change initialization that will need to be done across all address lookup pages, and address manual pages was being done in the specific operator page.

This change creates an initializer in each of the base address form objects, so that only initialization which is specific to being for the the operator is left in the operator address form objects.