Remove the duplication in the operator address forms
With the previous work to remove duplication and make things consistent across the address lookup and manual base forms, it left the operator address lookup and manual forms being almost exactly alike.

Therefore to reduce duplication, this adds a module to hold the duplicated code, which the 2 form objects now just include.