Remove need for 2 addresses attributes
Essentially both addresses and temp_addresses did the same thing but in different contexts.

When the form was backing the view template temp_addresses held the results found that the user would then be expected to select from.

When submitting addresses held the updated collection of address which will be assigned to the enrollment prior to saving it. Again its just temporarily holding the addresses so there was no reason why temp_addresses couldn't be reused in this context.