Add contact manual address page
This adds the manual address entry page for the WEX contact copying the pattern for manual operator address.

It also resolves an issue found with having multiple manual address entries in the shared manual address view partial. The shared partial included the link to change the postcode, which needs to be dynamic depending on the address we are working with. However prior to this change we had it set to the operator postcode lookup page.

Following the pattern set in the address lookup views, we simply moved this section out of the shared manual address partial and into the contact and operator views.