Add site grid reference page (#18)
This adds the site grid reference page to waste exemptions. This does not have an equivalent in Waste Carriers so we have taken the technical design from the waste exemptions model, but the UI design and content from Flood Risk Activity Exemptions (FRAE), which came after WEX and iterated it's site grid reference page.

This also brings in the [os_map_ref](https://github.com/DEFRA/os-map-ref) gem. Similar functionality was built into WEX, but then moved to this gem for FRAE. We use it to convert a National Grid Reference to an x & y (easting & northing) so we can then query the area lookup WFS service to determine the EA area. It also helps us validate that a grid reference is valid.