Change site description columns to be text (#19)
The next change we were about to work on was selecting the exemptions. One of the things we need prior to that is to import and store the exemptions to choose, and as we began work we were reminded there is a 'text' type in PostgreSql (we need to hold a description for the exemptions).

We then had a "doh!" moment when we realised that this would be a better way to hold the site description, as its more expressive of the type of information be held (multiple lines of text rather than a single text value) and gives us more flexitibility in the future.