Add select waste exemptions page
This adds the page where users can select the exemptions they wish to register. Not only do we not have a matching page in WCR, with the use of tabs in WEX, we don't really have a straight up design on which to allow users to select multiple options from a list of checkboxes.

So we have stuck with our standard pattern for all other pages, but when it comes to the logic in backing form object, this is a first.

Additional changes to support this page include

- adding exemption model and table
- adding the enrollment_exemption model and table (join table between enrollment and exemptions)
- exemptions.csv which holds all data on the exemptions that can be selected
- adding new seed task to import data from exemptions.csv into the database

It should be noted that we have gone with a `has_many :exemptions, through: :enrollment_exemptions` association between enrollments and exemptions from the off. We could have dropped the `through:` modifier but we already know that we will need to store additional information against the join model e.g. its state (ACTIVE, EXPIRED, REVOKED etc), comments, etc. Hence we wanted a model from the start which we can begin to build on, rather than trying to add it in at a later date.