Add check your answers page (#21)
The design and method for displaying previously entered answers is lifted straight from Waste Carriers. However as the content is different it has been heavily edited to match the Waste Exemptions journey.

There has been one slight tweak to the design as well. When it comes to referencing the locale content, rather than generic subheadings and fields from different sections aligned together, it has been broken down into sections.

This should hopefully make aligning the content and layout a little easier when changes are made in the future.