Add site postcode page
This is following our pattern for postcode search pages in the app, with one slight tweak.

Moving from the site grid reference to the site postcode is an optional step, and not as a result of an error. Also it does not infer we want to enter the address manually.

Hence we have added a `skip_to_address` method to the controller, and event in the workflow to better reflect this.