Add site address lookup page (#24)
This adds the page to select the site address based on results of a postcode lookup. It's essentially a duplicate of what we have already done for the operator and contact address.

To that end, we have also been a bit cheeky and added the code in `CanChangeWorkflowStatus` (though commented it out) to handle manual address entry, as it was easier to work out the transitions based on where we want to get to, rather than via an interim step.

They will be uncommented in a subsquent commit.