Add site manual address entry page (#25)
Follows the pattern set for manual address entry, and copies what we have done for the operator and contact.

In addition it includes a new feature specific to the site address, where we use the OS Places postcode lookup to determine the x & y (easting and northing) for a manually entered an address.

Bringing in the current design for address entry (originated in FRAE) makes it less likely that a user can get an unknown postcode in. With this additional change we can move a step closer to 100% of addresses having an X & Y, and therefore our ability to determine the EA Admin area they belong to.