Add declaration page
Using the WCR declaration as a model, we implement it here but update the content to match WEX's declaration page.

We also fix an issue found with the check details page, caused by the fact that `is_a_farm` and `on_a_farm` are stored as boolean's, though when we validate them in their own forms the params are passed as strings (hence the confusion around what validation to use).