Add registration complete page
Taking the logic of Waste Carriers we have updated the content to match the waste exemptions registration complete page.

Part of the changes require us for the first time to mark state of the exemption (active, ceased, revoked, expired etc) plus the dates it was registered and will expire. We also need to handle generating the reference number for the enrollment.

As with Waste Carriers, wherever we have to record and transition state in an entity we use a state machine. Hence the introduction of `can_change_exemption_status` in this change.