Rename Enrollment to registration
Enrollment is a name inherited from the previous WEX project, when the service was struggling to come up with a generic term for what the entity is prior to being an actual registration.

The Service owner at the time vetoed **Application**, which didn't leave us with many choices.

As time has gone on anyone new introduced to the project needs the history retelling to explain why its **Enrollment** and not **Registration**.

With hindsight we can see we refer to it in our content as a registration, and behind the scenes the most important semantic is that the dev team understands what an object is and for.

Hence whilst we have a chance with this re-write we are taking the opportunity to replace 'Enrollment' with 'Registration' throughout the project. This has meant changing previous migrations rather than creating new ones. However at this time this is a single dev effort so no one would be effected by this.