Refactor registration to transient_registration (#32)
Following a review the dev team came to a clearer idea of the meaning of `transient`. The term comes from Waste Carriers where during the renewal process we create a copy of the registration and refer to it as a `transient_registration` (both in the database and in the code).

Our definition is any record in the process of being changed. So for Waste carriers its a registration being renewed, and therefore edits are being made. Until the renewal process is complete we want to keep a copy of the original and the edited, hence 'transient'.

We believe the same will apply to a registration in Waste Exemptions. Until the registration process is complete, there is no 'registration'. And we know in the future we will need to build in the ability to make edits to a registration.

So we are refactoring the engine at this stage such that what we previously referred to as a registration is now the transient registration. When the registration process is complete, we convert the transient registration to an actual registration, and delete the transient.

This gives us a clear separation between the 2, and gives us the flexibility to build in editing in the future.