Add partners page (#39)
This is the page that lists all partners currently added and allows a user to add another, or delete an existing one.

It also includes the validations to ensure that the user enters at least 2 partners before they can proceed.

To keep both WEX and WCR consistent, it borrows heavily from Waste carriers which has the concept of different types of 'people'. Because of this we get a lot of code for handling other business types as well, even though WEX only cares about partners.

This just means we are better able to make changes in how we grab a list of 'people' in both services at the same time in the same way.