Send confirmation email when registration complete (#40)
In all our services when a user gets to the end of the journey we acknowledge it with an end page and an email. We already have the end page, this change adds the confirmation email.

Again code from the WCR renewals service has formed the basis of the send email functionality, and like the WCR we send the email on the completion service object.

A key distinction here is that we may be required to send multiple emails. In the existing service an email is sent to both the applicant and contact email addresses if they are different.

So to remain consistent we also have updated the end page to display an appropriate message if 2 emails are sent rather than just one.