Generate registration pdf & attach to conf. email
One thing both users and the internal support team have been asking for is to have a PDF copy of the registration automatically attached to the confirmation email we send.

With basing the WEX engine on the WCR one, we have been able to quickly migrate its PDF generation functionality, and used its template as a basis for one for WEX.

This change implements the functionality to generate a PDF and attach it to an email.