Implement configuration in the engine
Prior to this change configuration was essentially reading env vars directly, or assuming that the host application had set something in its application.rb and made it available via `Rails.configuration`.

As we started to look at adding our missing tests, we realised that we'd have to adopt this same pattern in our dummy app if we wanted this to work.

Ideally we want never to have to touch the code generated for the dummy app, hence we went looking for an alternative.

https://guides.rubyonrails.org/engines.html#configuring-an-engine details how to go about providing configuration for an engine, and this not only solved the test problem but also felt like a better approach.

Hence these changes implement the mechanism to allow a host application to pass in config values, and updates the code in the engine to use them rather than `Rails.configuration`.