Improve exemptions dates in PDF certificate
Prior to this change the expires on date for each exemption was displayed in the formatt YYYY-MM-DD. To be consistent with other dates in the certificate we wanted to change this to be 'January 22 2019'.

The longer format would have forced each exemption to cover 2 lines, so we also tweeked the formatting of the exemption details section and how the date is shown in each row.