Remove notice icon from declaration page
We were having trouble aligning the icon to some text in the declaration page, when it was highlighted that we don't use any icons in the Waste carriers renewals declaration page.

Therefore to keep the services consistent this change removes the image (and therefore the problem) from the declaration page.