Fix site validation in check answers page (#46)
Because of the way we had the validation configured previously it was only letting us enter a site location using grid reference and description.

This change adds a conditional to the validation i.e. if the site was set using a grid reference use one set of validation, else use the other.