Convert to block style configuration
We have a known issue with the regex in the companies house validation and as we copied this from Waste Carriers it means the same issue is there. Rather than simply updating both, we are building a shared gem for validation that can then be used by both services.

As part of building the [Defra Ruby Validators gem](https://github.comn/DEFRA/defra-ruby-validators) we needed to provide it with the companies house API key.

We didn't want to just rely on having the gem read env vars directly so we followed a common pattern of the host app using a configuration block to set the gem's configuration

```ruby
DefraRubyValidators.configure do |configuration|
  configuration.companies_house_api_key = ENV["COMPANIES_HOUSE_API_KEY"]
end
```

Prior to this change the engine followed a slightly different pattern of using module accessors that are called directly

```ruby
WasteExemptionsEngine.companies_house_api_key = ENV["COMPANIES_HOUSE_API_KEY"]
```

The problem we faced is that the actual implementation of the gem would be in this engine, which in turn would be hosted in the target application. And we need the host app to set configuration values, which are passed all the way down to the gem.

When first implementing we realised that the engine's initialisation of the gem was being called before the apps initialisation of the engine.

We found that if we converted to using the block configuration pattern for the engine as well as the gem, we could also pass values down to the gem when the engine was updated. It also made configuration consistent.

So this change is in preparation for implementation the new validators gem 😁