Fix broken rspec test environment
Though we had setup rspec plus some other dependencies when we created the engine, as we have been 'naughty' and up till now not added any actual tests we hadn't realised it was actually broken.

With no tests, rspec was not truly spinning up so we were not spotting that references to dependencies like database cleaner would cause an error, and that in the Travis CI environment the migrations had not been run so nothing would actually work there.

So these changes, including an initial test which touches the database and makes use of Factorybot should resolve the issues with the setup and ensure the tests we'll be adding will run.

For future reference the patterns we have tried to stick to with these changes are

**Gemspec only dependencies**

There is some confusion as to where to declare dependencies in a gem/engine as you have both a `Gemfile` and a `gemspec`. Where consensus lies is adopting the advice of Yehuda Katz, which is that [all dependencies should be in the gemspec](https://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/).

Essentially the `Gemfile` should be used when working on a gem. So referencing locally gems or Github repo's would be done here. But once the gem is ready for release, the finalised dependencies should be added to the gemspec and the `Gemfile` returned to its vanilla state.

**Test dependency initializers**

Where we have dependencies when running tests, rather than mix their setup with the standard config in `rails_helper.rb` and `spec_helper.rb` create an initializer under `spec/support`. This should help make it clear what setup is specific to what dependent gem.

**Spec first, rails after**

The idea of the separation of the `rails_helper.rb` and the `spec_helper.rb` is to make a distinction between setup needed when running a rails based tests, and non-rails tests. So if testing a file in `lib/` which doesn't depend on Rails you may get a performance boost by just requiring `spec_helper.rb` in its spec.

In practise there is a lot of overlap between the 2, for example the line which automatically requires anything in `spec/support` needs to be in `rails_helper` because Factorybot needs to be initialised and configured after rails has been setup to avoid an error.

However if possible, if something doesn't require rails, aim to reference it in `spec_helper.rb` first.