Fix issues with company no. validation
It was spotted during the testing of the new service that if you entered a registration number using lowercase characters, or added any whitespace then the check against companies house would fail.

The registration form should handle this on behalf of users rather than returning an error and expecting them to correct it. The changes included here will resolve both problems.