Make choose exemptions checkboxes more accessible
It was noted on the existing service that the accessibility of the checkboxes could be improved, specifically how they appear to screen readers.

This also applied to this version. Prior to this change the screen reader would just read the code for the checkbox. So a user tabbing through would just hear "U1. U2. U3. Selected U3".

With this change they will now also hear the summary other users can see visually on the page e.g. "U1 Use of waste in construction. U2 Use of baled end-of-life tyres in construction".