Fix database cleaner config
When working on PR #53 found we found we were struggling to get the tests to run without PG errors about violation of unique keys popping up.

Some digging found the database was not actually being cleaned before the tests started. A review of [DatabaseCleaner's guidance](https://github.com/DatabaseCleaner/database_cleaner#rspec-example) suggested a different config setup to what we have, and implementing it resolved all our issues with the test.

Hence this change to fix the config.