Implement defra ruby validators (company no.) (#51)
A core driver of building a new version of the Waste Exemptions Service based on the Waste Carriers Renewals pattern was to resolve a long list of known issues with the existing service.

One of them was that it would treat company registration numbers like IP00141R as invalid. However even after we migrated the new company registration number validation from WCR it still would not accept a number formatted as XX00000X.

Turns out the regex we had implemented in WCR renewals, had made the same mistake as previous implementations and missed these kinds of company registration numbers.

So now we were faced with correcting the flawed regex in 2 different places. Because of this we instead created [Defra Ruby Validators](https://github.com/DEFRA/defra-ruby-validators) so we could define validations in one place, and share them across our Rails services.

Having built the new gem, this change implements it in the engine.