Fix our simplecov configuration (#57)
We always drop simplecov into our projects to give us our test coverage. We did this as usual in this project and were a little surprised to see very high test coverage even though we know we are very few tests 😳.

As we have started adding tests the total coverage has started to fluctuate quite drastically so we took a deeper look at our test coverage.

What we found is that

- our dummy app's code files were being included, and reported as having 100% test coverage which inflated our overall figure
- simplecov was only reporting test coverage for files 'touched' by the existing unit tests

Clearly this is wrong and not giving us the accurate figure we need to make a proper assessment about our unit test coverage.

Therefore this change is intended to fix the issue. It follows our adopted pattern of configuring test dependencies in separate files, but ensures it still fulfills simplecov's requirement of being the first thing required. It also ensures it doesn't get included twice.

Other than that it makes sure we exclude files from the coverage that are irrelevant, whilst ensuring that all files in the project are included and not just the ones touched by unit tests.