Handle refresh on the completion page (#56)
https://eaflood.atlassian.net/browse/RUBY-23

It was noted in testing that if you hit refresh on the last page of the journey you would get an error.

The problem is that between the declaration page being submitted and the end (completion) page being shown the transient registration has passed through the `RegistrationCompletionService`. This converts the transient registration to a registration, and deletes the transient.

So if a user hits refresh they are essentially requesting they wish to see a transient registration that no longer exists.

In consultation with our QA and the Product Manager we have agreed that if a user makes a request for a transient registration that does not exist (the **token** is not recognised) they will be redirected to the root page of the service.

This change ensures that is what happens in this scenario. For example it was noted that even though we previously requested a path without token in `FormsController.form_path` in the code, if you debugged and checked what happens the helper automatically adds the existing `params[:token]` to the path anyway. By specifiying the path rather than using the helpers we are making it clear what our intention is.

We have also added a basic controller test to cover our expected behaviour in this scenario.