Remove rspec use transaction fixtures config
When aligning the engine's rspec setup with the back-office-ta project spotted that we had the line

```ruby
config.use_transactional_fixtures = false
```

in `rails_helper.rb`. We didn't have this in the back-office-ta project, and as we are using Databasecleaner it doesn't make sense to also have this line confusing issues. Hence this change removes it.