Inherit authorisation/authentication preferences from host (#62)
Improved testing during https://github.com/DEFRA/waste-exemptions-back-office-ta/pull/21 revealed that the mounted engine wasn't using any of the auth checks provided by Devise in the host app - an unauthenticated user couldn't access routes from the host app, but could access ones in the mounted engine.

This also applied to checks for abilities as well.

The front-office doesn't use Devise or CanCanCan to manage authorisation or authentication, but the engine should be able to handle it when the host app specifies it. Inheriting the host's ApplicationController allows this.