Fix postcode form validation (#61)
https://eaflood.atlassian.net/browse/RUBY-83

The postcode forms fail to correctly handle postcodes that match zero addresses. The expected behaviour is that the form validation displays an error asking the user to either correct the postcode or manually enter an address. This matches the behaviour on Waste Carriers.

This work includes additional changes to the automated tests both to setup what is required to test the postcode form but also to support upcoming work on the automated tests.