Refactor AddressValidator (#65)
In an effort to remove duplication in the validators, we would like to move duplicated code into the ValidatesPresence concern and BaseValidator class.

Future refactoring efforts will include the creation of the ValidatesLength concern.

This PR is a sample of the intended refactoring. We decided it would be easier to review the changes if they are applied to a single validator before applying the pattern to the rest of the validators.