Fix transpose farm answers in check answers page (#60)
https://eaflood.atlassian.net/browse/RUBY-99

An issue with the current version of the service is that the answers to whether the waste activity will

- take place on a farm
- the person responsible is a farmer

get transposed in the check your answers screen. When we migrated the functionality to this engine, we also inadvertently copied this same transposing issue.

The root cause is the bad naming of the routes. Prior to this change they were

- `is_a_farm` meant is the activity taking place on a farm
- `on_a_farm` meant is the person responsible a farmer

It's easy to see why with names like this the information gets transposed. So this change resolves the root problem by renaming the routes, and all related objects (controllers, forms, views, column names etc), and ensures the correct info is displayed with the correct title in the check your answers page.