Alias transient_registration associations to non-transient names
Came out of https://eaflood.atlassian.net/browse/RUBY-154

I ran into issues in the back office when trying to reuse code for displaying registrations and transient_registrations, as all their associations, while very similar, had different names.

This should allow us to have a `result` which is a registration or a transient_registration, and be able to call `result.addresses` or `result.people` to get the desired result.